In branch predictor history, record predTaken based on result after BTB.
Before this change, when the branch predictor predicts that the branch is
taken, but no BTB entry is found the branch predictor history will say
it is taken. Then, if the branch is committed as not taken, the branch
predictor will be told that that it was taken, since the prediction
in the history is not updated.

Change-Id: Ie2bd8bf35d635c1cd267e586b39bddb6723ddd92